pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "haldxai"
dynamic = ["version"]
description = "Healthy Aging and Longevity Discovery AI Platform"
readme = "README.md"
license = {file = "LICENSE"}
authors = [
    {name = "HALDxAI Development Team", email = "haldxai@example.com"},
]
maintainers = [
    {name = "HALDxAI Development Team", email = "haldxai@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
]
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scipy>=1.7.0",
    "scikit-learn>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "neo4j>=5.0.0",
    "sqlalchemy>=1.4.0",
    "spacy>=3.4.0",
    "transformers>=4.20.0",
    "torch>=1.12.0",
    "openai>=1.0.0",
    "matplotlib>=3.5.0",
    "seaborn>=0.11.0",
    "plotly>=5.0.0",
    "networkx>=2.8.0",
    "mpl-chord-diagram>=0.2.0",
    "pyyaml>=6.0",
    "python-dotenv>=0.19.0",
    "tqdm>=4.64.0",
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "fastapi>=0.85.0",
    "uvicorn>=0.18.0",
    "typer>=0.6.0",
    "click>=8.0.0",
    "rich>=12.0.0",
    "pathlib2>=2.3.0",
    "typing-extensions>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "mypy>=0.971",
    "pre-commit>=2.20.0",
]
jupyter = [
    "jupyter>=1.0.0",
    "ipykernel>=6.0.0",
    "ipywidgets>=7.6.0",
]
scientific = [
    "biopython>=1.79",
    "rdkit>=2022.3.5",
]

[project.scripts]
haldxai = "haldxai.cli:main"

[project.urls]
Homepage = "https://github.com/your-org/HALDxAI-Repository"
Documentation = "https://haldxai.readthedocs.io/"
Repository = "https://github.com/your-org/HALDxAI-Repository"
"Bug Tracker" = "https://github.com/your-org/HALDxAI-Repository/issues"

[tool.setuptools]
packages = ["haldxai"]

[tool.setuptools.package-data]
haldxai = [
    "configs/*.yaml",
    "data/*.json",
    "templates/*.html",
]

[tool.setuptools_scm]
write_to = "haldxai/_version.py"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["haldxai"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "networkx.*",
    "matplotlib.*",
    "seaborn.*",
    "plotly.*",
    "sklearn.*",
    "spacy.*",
    "transformers.*",
    "torch.*",
    "neo4j.*",
    "psycopg2.*",
    "biopython.*",
    "rdkit.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["haldxai"]
omit = [
    "*/tests/*",
    "*/test_*",
    "setup.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]